+++ packages/core/src/docdb/explorer/dbInstanceNode.ts
@@ -7,7 +7,7 @@ import * as vscode from 'vscode'
 import { inspect } from 'util'
 import { AWSTreeNodeBase } from '../../shared/treeview/nodes/awsTreeNodeBase'
 import { DBInstance } from '../../shared/clients/docdbClient'
-import { DocDBContext, DocDBNodeContext } from './docdbNode'
+import { DocDBContext, DocDBNodeContext } from './docdbContext'
 import { DBClusterNode } from './dbClusterNode'
 import { ModifyDBInstanceMessage } from '@aws-sdk/client-docdb'
 import { waitUntil } from '../../shared'
